internal/envoy: remove RoutePrefix and RouteRegex
Updates #1351

Prior to #1882 the envoy package offered two ways to construct a
Prefix/Regex match -- the first was a direct call to
RoutePrefix/RouteRegex which took a string parameter, the second was
via RouteMatch which took a dag.Route object. #1882 refactored
RouteMatch on top of RoutePrefix/RouteRegex which consolidated on one
method of constructing a regex match, #1913 later took advatage of this
to update all uses of regex to safe regex.

However there still existed two ways to construct a matcher, one via a
direct call to RoutePrefix/RouteRegex the other via RouteMatch. It turns
out that all the test code used the former, and all the production code
used the latter, implying that RoutePrefix/RouteRegex were test helpers.

This PR removes RouteRegex/RoutePrefix and replaces their calls in test
code with helpers that construct the respective *dag.Route object and
passes that to RouteMatch.

Signed-off-by: Dave Cheney <dave@cheney.net>