internal/envoy: remove clusterDefault
In bae8971f8 the logic for setting Cluster.Name, AltStatName,
CommonLBConfig and LbPolicy was refactored into clusterDefaults and the
results from the latter merged with a test fixture during tests. This
made the TestCluster fixture smaller at the cost of the reader of the
test having to understand that the fixture as presented was not complete
-- it was merged with clusterDefaults before comparison.

For the saving of two lines I don't think clusterDefaults paid for
itself, especially as the clusterDefaults + proto.Merge behaviour was
not used in code, only in tests, where as in envoy.Cluster the result
from clusterDefaults was overwritten, not merged.

This PR rolls clusterDefaults back into envoy.Cluster adjusting the
cluster_test fixture to match.

Signed-off-by: Dave Cheney <dave@cheney.net>