internal/featuretest: add test to assert tlspassthrough + permit inseure works
Updates #1945
Updates #1954

Clean up the existing tcp proxy feature tests to assert that no
listeners are opened unexpectedly. Then add a test to assert that the
pattern of exposing a service both via tcpproxy and port 80 (permit
insecure) works

There are several bugs that have fallen out of this

- #1951 - Contour accepts a httpproxy with a missing service without
marking the httpproxy invalid.
- #1952 - httpproxy's tcpproxy plus permit insecure does not require the
service to be marked as permit insecure.
- #1954 - route should not be present on ingress_https if tcpproxy is
inuse.

Both of these should be fixed, but lets lock in the current behaviour.

Signed-off-by: Dave Cheney <dave@cheney.net>