internal/envoy: bump the max Envoy regex program size
There's no obvious way to know what a reasonable regex program
size is (and that's not really what the RE2 program size number
means).  However, Envoy makes us send something here and if it's
too low, it will NACK the update and Contour will not handle that
gracefully. So, we send a large number that should be enough for
all reasonable regex cases.

This fixes #2043.

Signed-off-by: James Peach <jpeach@vmware.com>