internal/grpc: improve xDS logging
When Contour logs the close of a xDS stream, the log entry doesn't
include as much information as it could because it is captured before
and requests are made. Rather than defering the log, use a helper
function to write the log entry once it has been decorated with all
its fields.

Signed-off-by: James Peach <jpeach@vmware.com>