docs: add generated API reference
Add some site scripts to generate API reference documentation for
the Contour CRDs. Unfortunately, we need to vendor the Go templates
from the documentation generator in order to add enough CSS to make
the results attractive.

Note, that if running the generator has the side-effect of updating
Go modules. These changes need to be reverted afterwards.

Signed-off-by: James Peach <jpeach@vmware.com>