internal/metrics: Add healthz metrics
Fixes #1560

Signed-off-by: Nathan Bass <ingcr3at1on@gmail.com>

Tests cover total and successful requests but we need to be able to pass in
a _different_ fake kubernetes.Interface for the kubernetes.Clientset to trigger
the unsuccessful flow. I don't think anyone with projectcontour wants to maintain
a separate copy of the kubernetes.Interface so at the moment that is left out.

Changes the kubernetes.Clientset expectation to the kubernetes.Inteface in
registerHealthCheck because otherwise we can't pass in a fake.Clientset, this means
the call to ServerVersion has to be made on teh Discovery object instead of directly
on the Clientset as ServerVersion is not otherwise defined on the kubernetes.Interface.