cmd/contour: improve Kubernetes client error message
If Contour failes to create a Kubernetes API clien on startup,
add context to the message to make it clear what kind of problem
happened.

For example, if you are running a local Contour and don't have a
current Kubernetes context, Contour would previously exit with the
message "invalid configuration: no configuration has been provided".
This doesn't make it clear the problem is the Kubernetes config,
not the Contour config.

Signed-off-by: James Peach <jpeach@vmware.com>