site: stop linking absent document versions
If a new versioned document is added, it may not be present in all
versions of the documentation. In this case, the version selection
button should not generate a link to an absent version of the
document.

The fix is to scan the site pages before generating the link to
ensure that the target version of the document is actually present.
If the target version isn't present, then we split the the page URL
on the version number token, and then append the latest version.
This has the result of linking back to the root of the latest
versioned documentation tree.

This fixes #2085.

Signed-off-by: James Peach <jpeach@vmware.com>