internal/contour: refactor metrics test for accurate line numbers
When a test fails, the reported line number is in the subtest, and
you have to search for the tesst case name to find the test case
that had a problem. Refactor the test code into a local function
so we can preserve the attributed of table-driven tests but aldo
get accurate reporting of line numbers for failures.

Signed-off-by: James Peach <jpeach@vmware.com>