internal/envoy: raise SafeRegex limit to 1048576
Updates #2240

Raise the SafeRegex size limit from 1,000 to 1048576. There is no evidence
that this number is sufficient for all possible regex patterns, thus the
limit represents the "no limit" limit because it is currently not
possible for envoy to reject a regex entry in a way that Contour can
trace back to the original input.

Signed-off-by: Dave Cheney <dave@cheney.net>