internal/contour: record EventHandler operation metrics
Fixes #2235

Introduce a new handler in the EventHandler chain which emits prometheus
metrics for the various API operations and kinds observed from informers
which sink to the EventHandler chain.

Inserting the EventRecorded between the DynamicClientHandler and the
EventHandler will make this PR easier to backport to Contour versions
which don't use the dynamic client.

Signed-off-by: Dave Cheney <dave@cheney.net>