internal/k8s: move dynamic client translation to k8s package
Updates #2235

Contour 1.2.x is moving to the k8s dynamic client for deserialising
Contour's CRD objects. To make #2244 easier to backport to Contour's 1.1
branch, which does not support the dynamic client, break this logic out
into its own ResourceEventHandler wrapper which returns the current
contour.EventHandler logic to Contour 1.1 spec.

As a side effect, we only need to use the DynamicClientHandler wrapper
in cases where we know the resources are coming from the DynamicClient.
Specifically we don't need to use this path for Core.v1 and Extension
objects. Also, we don't need to use the DynamicClientHandler wrapper in
e2e/feature tests as there is no k8s API server connected to those tests
-- we construct the final k8s object as a fixture and call
contour.EventHandler.OnAdd/Update/Delete directly.

Signed-off-by: Dave Cheney <dave@cheney.net>