internal/contour: remove holdoffmaxdelay forced update
Updates #2275

The holdoff logic is subtle, and under tested. To address both of these
issues I'm going to make the fix for #2275 in several stages so we have
a chance to git bisect.

This first change removes the explicit force code path when
holdoffmaxdelay is exceeded. The logic remains the same, if
holdoffmaxdelay is exceeded a delayed update will occur, however after
this change rather than their being a specific code path for the forced
update we set the timer delay to zero, which will make the <-pending
channel ready shortly afterwards.

Strictly speaking a stream of <-op updates can trump a <-pending which
can add further delay to rebuilding the dag but in practice this is
unlikely to occur indefinitely;

1. When more than one channel is selectable, the runtime chooses between
them pseudo randomly.
2. If <-op updates occur continually it is in the users' interest that
the DAG rebuild is delayed, this is the point of 2275.

A future PR will refactor the delay calculation so that we can test it
in isolation.

Signed-off-by: Dave Cheney <dave@cheney.net>