internal/k8s: simplify DynamicConverter
Updates #403

Currently in cmd/contour.doServe we have three ResourceEventHandler
types defined;

- eventHandler, which is the internal/contour.EventHandler connected to
the dag
- eventRecorder, which wraps eventHandler to add metrics
- dynamicHandler, which wraps eventRecorder to handle dynamic objects.

These three objects are all in scope and used to register with various
informers in a fragile pattern; eventHandler needs to be in scope to
be registed with the group later in the function, eventRecorder is
registered with the k8s core types, and dynamicHandler with the CRD
types. This is all rather fragile.

This PR prepares to make the DynamicHandler the only entry point for all
k8s types by refactoring the DynamicHandler to make accepting non
dynamic types less unexpected.

Signed-off-by: Dave Cheney <dave@cheney.net>