internal/contour: enforce SNI name bindings
Previously, Contour generates a single RouteConfiguration named
`ingress_https` that contained all the virtual hosts that were being
served over HTTPS. This meant that if the client request SNI name "A",
it could send a HTTP request for virtual host "B" and it would work.
This behaviour is generally unexpected and can violate security invariants
that we want to enforce, e.g. protecting virtual hosts with client
certificate validation.

The fix is to generate a unique route configuration for each virtual host
and match them to each unique SNI name. This means that once the SNI name
is matched, all the routes from the attached route aonfiguration belong
to the same virtual host. This enforces a one to one correspondence
between SNI names and virtual hosts.

After this, the well-known "ingress_https" route configuration no longer
exists, except for TCP proxying.

This fixes #2347.

Signed-off-by: James Peach <jpeach@vmware.com>