build: support older git versions to detect current branch (#2415)
Ubuntu 18.04 has git version 2.17.1 which doesn't have the --show-current
flag available (It's in a newer version of git). This causes builds to
throw an error when the makefile looks for the current branch. Go back to
the old method of parsing `git branch` output to get the current branch.

This fixes #2414.

Signed-off-by: James Peach <jpeach@vmware.com>