Add github label automation.
Pull in a Github labels configuration file based on the Kubernetes labels
configuration. This enables a single source of truth for labels within
the projectcontour organization, and lets us document the purpose of
each label.

Add a proposed wrapper script that can be run in a scheduled Github
Action to periodically sync the labels.

Update the process documentation to show the labels and give advice on
how they should be used in the triage process.

Signed-off-by: James Peach <jpeach@vmware.com>