doc: improve ambiguous upstream TLS example
The text that attempts to describe using annotations to enable upstream
TLS doesn't match the example properly, and the example isn't very clear.

Update both the HTTPProxy and the annotations documentation for clarity,
and add a better example that demonstrates how to apply the protocol
annotation.

This fixes #2457.
This updates #2337.

Signed-off-by: James Peach <jpeach@vmware.com>