Adds support for updating HTTPProxy addresses.
This is step 2 in a three phase plan.

Adds the `status.loadBalancer` struct to HTTPProxy.

Also, updates the StatusAddressUpdater with HTTPProxy support.
This will probably be the last type added there for a while.

I experimented with adding a `kubectl get` column for the addresses, but the
multi-value nature of the loadBalancer.ingress slice is not accessible via a simple
jsonPath expression, so we can't get all the values without special logic (which I
assume Ingress has).

Address details are available via `kubectl describe` from the command line.

Updates #2371.

Signed-off-by: Nick Young <ynick@vmware.com>