cmd/contour: replace current xds server with go-control-plane impl
Fixes #2134 by implementing the Envoy go-control-plane to replace Contour's current custom xDS gRPC server.

The change utilizes snapshots as a way to represent a versioned point in time representation of the xDS
resources (RDS, CDS, EDS, LDS, SDS). When the dag is rebuilt or an endpoint changes, a new snapshot is
created with the updated caches for each xDS resource type.

Signed-off-by: Steve Sloka <slokas@vmware.com>