cmd/contour: support defaulting HTTP protocol versions (#2622)
Add Contour config file for defaulting the HTTP protocol version
accepted by the proxy listener. This principally applies to HTTPS
listeners, where HTTP/1.1 and HTTP/2 can be accepted on the same
listener.

If no configuration is specified, all HTTP protocol versions are
automatically detected and supported. Otherwise, HTTP/1.1 and HTTP/2
can be individually toggled. There is no support for per-resource
configuration in this change, but it can easily be supported if
necessary.

The config file syntax looks like this:

```
default-http-versions:
- http/1.1
- http/2
```

This updates #2619.

Signed-off-by: James Peach <jpeach@vmware.com>