apis/projectcontour: rename the HTTPProxy Status type
Rename the `Status` struct to `HTTPProxyStatus`. This means that the
type name won't conflict when we bring new CRDs into the `v1` namespace,
and makes the nomenclature consistent with established CRD patterns.

This is a no-op for API users, since the field name doesn't change,
but is an incompatible source change for projects using the Go types.

Signed-off-by: James Peach <jpeach@vmware.com>