Add conditions to HTTPProxy
Updates #2495.

This adds a new `DetailedCondition` type to HTTPProxy, that allows for sub-conditions.

This allows Contour to have a `Valid` condition, that has further information about
how that conditions `status` has been reached, or in other words, to expose more
than one error or warning about a HTTPProxy to the user.

Also adds the Conditions block to TLSCertificateDelegation.

Signed-off-by: Nick Young <ynick@vmware.com>