apis/projectcontour: add ExtensionService API (#2719)
Add the `ExtensionService` API resource definition. This will
be present internally to Contour and to programs that import the
Contour `apis` package, but is not installed as a CRD unless the
`ENABLE_EXTENSIONS_CRD=Y` make variable is specified. This option
will be removed once the full `ExtensionService` API is implemented.

This fixes #2710.

Signed-off-by: James Peach <jpeach@vmware.com>