cmd/contour: refactor shutdown-manager to initiate shutdown using an Exec command
This changes the preStop lifecycle hook to utilize a command instead of an http request.
The /shutdown http endpoint now only returns when the pod is ready to be terminated, and
no longer initiates the Envoy shutdown sequence.

The Envoy container will block on its preStop hook until the /shutdown endpoint returns.
That endpoint, implemented in the shutdown-manager, will utilize a file messaging pattern
to know when it is safe for Envoy to terminate by checking for the existence of a file
in the filesystem.

Signed-off-by: Steve Sloka <slokas@vmware.com>