internal/contour: simplify the endpoints translator
Breaking EndpointsTranslator into two parts doesn't provide any benefits
and we can make the code easier to follow by removing all the indirection.

The `recomputeClusterLoadAssignment` member function was hoisted out
to be a free function so that the `EndpointsTranslator` lock policy is
clearer. This function doesn't need to hold any `EndpointsTranslator`
locks or operate on any internal state.

Signed-off-by: James Peach <jpeach@vmware.com>