testsuite: add a fallback certificate test
Add a simple integration test to ensure that using the fallback
certificate works as expected. This requires that Contour is deployed
with a valid fallback certificate. For integration testing, we can use
cert-manager to pre-configure that, but this approach won't work well
for large numbers of configuration permutations, so we should revisit
it relatively soon.

Signed-off-by: James Peach <jpeach@vmware.com>