internal/dag: improve status errors for invalid TLS
When validating an HTTPProxy's TLS configuration, separately check
for cases where both secretName and passthrough are specified, and
for cases where neither secretName nor passthrough are specified.
Sets a corresponding HTTPProxy status message for both cases. Also
improves the status message for when TCPProxy is enabled, but TLS
is not enabled.

Signed-off-by: Steve Kriss <krisss@vmware.com>