internal/dag: stop embedding ServicePort in dag.Service
The Kubernetes `v1.ServicePort` is embedded in `dag.Service` as a pointer
field. There's no need to have a pointer here, since it is always assumed
to be non-nil, and using the type as an explicit field makes is easier to
see where it is used, and clarifies overlapping field (i.e. `Protocol`).

This updates #2769.

Signed-off-by: James Peach <jpeach@vmware.com>