cmd/contour: add a logger to the Builder (#2822)
The `Builder` should not use the `KubernetesCache` to log errors. This
gives the false impression that an error occurs in the resource caching
layer rather than the builder layer. For now, we just use a generic
"builder" context name, but as we work on splitting `Builder` up, we
can use separate HTTPProxy and Ingress builder context names.

Signed-off-by: James Peach <jpeach@vmware.com>