internal: rebuild endpoints based on the DAG services (#2805)
The EndpointsTranslator generates Envoy ClusterLoadAssignment resources
in a fixed configuration. Each service/port pair generates a separate
ClusterLoadAssignment, that is associated with an Envoy Cluster solely
by its matching name.

This change introduces a new ServiceCluster concept that represents a
single set of endpoints, weighted across multiple Kubernetes Services.
This allows the upcoming ext_auth support to weight traffic, even though
(by definition), it still uses a single Envoy Cluster. The weighted
balancing is done within the cluster.

The DAG Service object is updated to use a ServiceCluster rather than
referencing a Service directly. This is exposed to the EndpointsTranslator,
which is now a DAG observer. The EndpointsTranslator rebuilds the
ClusterLoadAssigments globally on a DAG rebuild, and incrementally
on an Endpoints update.

The only semantic change here is that Contour will no longer generate
Envou ClusterLoadAssignment resources for all Kubernetes Services. They
will only be generated for Services that are referenced by an Ingress
or HTTPProxy resource.

This updates #2769.

Signed-off-by: James Peach <jpeach@vmware.com>

Co-authored-by: Steve Sloka <slokas@vmware.com>