Second attempt at fixing knative-sandbox/net-contour#226
Replace the dynamic client Get (uncached) with a read from
the informer's lister cache.  This code path is evaluated
for every HTTP Proxy on each DAG rebuild, and the call to
the API Server on each results in a thundering herd that
triggers the built-in rate limiting in k8s.io/client-go.

With this change, the only API Server calls on DAG rebuilds
happen only when our cached copy of the HTTP Proxy shows
that we need to perform a status update.

Updates #2857

Signed-off-by: Matt Moore <mattmoor@vmware.com>