internal/dag: Change status mechanisms to be ready for Conditions
This commit updates a number of pieces in `status.go` to be ready for adding conditions.
In general, these are to make the names clearer and to keep space for us to add further
support for other objects later as well.

`StatusUpdate` is renamed to `ProxyStatusUpdate` as it's only used for HTTPProxy.
`ObjectStatusWriter` is renamed to `ProxyStatusWriter`, for the same reason.

`StatusWriter`'s status cache has been renamed to `proxyStatuses` from `statuses`, in
case we want to use it for storing other object updates later.

Short values have been updated to reflect the new struct names (`osw` changed to `psw`, etc.)

`SetInvalid` has been updated to `AddInvalidCondition`, with a new `conditionType`
argument, that is currently only used to add some extra text into the `description`
field of the HTTPProxy. This will mean that further PRs should not need to touch the
HTTPProxy Processor.

This is currently a WIP however, the tests have not been updated yet. Since we
moved to `assert.Equal` for errors, the errors messages must now match for
tests to pass.

Signed-off-by: Nick Young <ynick@vmware.com>