internal/dag: add support for ExternalService clusters (#2876)
Refactor the ExtensionService `.Services` field so that we have enough
information to build an EDS cluster load assignment. Dropping the use of
the HTTPProxy `Service` type means that the API aligns more closely with
the corresponding Envoy resources and we can build a single Envoy cluster
that balances across all the configured Kubernetes services. This wasn't
previously possible, since Envoy generally expects external services to
be a single cluster and has no provision for balancing across multiple
clusters in this part of its API.

Add a new `dag.Processor` implementation to build ExtensionService
objects in the DAG. These are root-level objects that generate Envoy
cluster resources. The `.Services` field emits a `dag.ServiceCluster`
object that generates an Envoy cluster load assignment resource to
balance traffic within the cluster.

This fixes #2713.

Signed-off-by: James Peach <jpeach@vmware.com>