internal/dag: Update how headers are processed for special 'Host' header
The 'Host" header value is special when applied to a
requestHeadersPolicy. When this is set it configures
the hostRewrite field in envoy to match the value of the
requestHeadersPolicy value field.

However, setting this value is only valid when applied to
a route in Envoy, it cannot be applied to a cluster, so
any user setting the "host" header field on a requestHeadersPolicy
on a Spec.Route.Service is invalid.

Fixes #2872

Signed-off-by: Steve Sloka <slokas@vmware.com>