internal: Add code for DAG status to emit Conditions (#2931)
This commit adds code for DAG status updates to emit Conditions
instead of updating the `currentStatus` and `description` fields
only. Instead, there is a `Valid` condition that indicates the same
information, but allows extra details to be added via SubConditions.

To do this, there's a new `status` package that handles status during
the DAG build, including a cache. This has been built with extensibility
towards ExtensionService and the service-apis types in mind, so we
should be able to add support for emitting Conditions on them all
as well reasonably easily.

After discussion on #2886, I've also locked in the behavior of the
additions to the `apis/projectcontour/v1` helpers with tests. I'll
do another pass after I finish out this work and add it to the auth
helpers as well.

Signed-off-by: Nick Young <ynick@vmware.com>