internal: update ExtensionService status conditions
Add status cache support for `ExtensionService` resources. This is built
on a more general cache interface, which ought to be extensible to more
types, but that whill need to wait for additional refactoring passes.

Adopt the status cache in the `ExtensionService` processor so that
conditions are correctly updates for `ExtensionService` resources.
The processor doesn't fail on errors, which allows multiple errors
to be published in the same condition.

This fixes #2874.

Signed-off-by: James Peach <jpeach@vmware.com>