internal/k8s: Ignore LastTransitionTime for HTTPProxy objects status equal check
Change how the comparison of the status of the object of HTTPProxy objects by
ignoring the LastTransitionTime which is always updates on each DAG rebuild
regardless if the status of object changed or not.

Not ignoring this causes each status to be updated each time since the objects
are always different for each DAG rebuild.

Fixes #2979

Signed-off-by: Steve Sloka <slokas@vmware.com>