pkg/config: export a package for config parameters (#3023)
Move the configuration file handling into a separate package. This
cleanly separates the configuration file from the contour internal
server details, making it clearer which struct fields are internal
and which are configuration (and therefore have compatibility concerns).

The configuration separation allows the configuration to be validated in
one place prior to setting up the Contour controller. This change modifies
the server start up to do that. This guarantees that the configuration
is which lets us remove subsequent error handling.

The new configuration package is exported so that other projects in the
Contour ecosystem can consume it. This means that the API here should
be considered (reasonably, but not absolutely) stable.

This fixes #2621.

Signed-off-by: James Peach <jpeach@vmware.com>