internal/xds: clone xDS protobufs before rewriting them
Each xDS resource caches the protoufs that it generates from DAG updates
and hands them out to the xDS consumers via a pointer reference. When
we rewrite these to xDS v3, we were modifying the cached values, which
would cause errors if we had to rewrite them again.

This updates #1898.

Signed-off-by: James Peach <jpeach@vmware.com>