Fix integration test failure
The auth integration test was finding a problem; that in v3
the HTTP router filter must be the last in the list. This was not the
case when external auth was enabled.

This commit updates the `AddFilter` logic to always move the Router
filter to the end of the filter chain, whenever any filter is added,
and updates the tests accordingly.

It is also technically possible, though a programmer error, to add
a second Router filter to the filter chain. Added a panic in this case,
as there is no way it's possible using the `AddFilter` method. This
catches a programmer error and fails fast and hard to tell that person
not to do that, and a test to lock in the behavior.

Signed-off-by: Nick Young <ynick@vmware.com>